Add route based code splitting
