+++ client/src/containers/Main/Main.js
@@ -10,12 +10,61 @@
 import React from "react";
 // objects match, location, history
 import { Switch, Route } from "react-router-dom";
+import Loadable from "react-loadable";
 
-import LandingPage from "../LandingPage/LandingPage";
-import Expences from "../Expences/Expences";
-import Register from "../../components/Register/Register";
-import Login from "../../components/Login/Login";
-import Info from "../../components/Profile/Info"; // change name
+import CircularProgress from "@material-ui/core/CircularProgress";
+
+// import LandingPage from "../LandingPage/LandingPage";
+// import Expences from "../Expences/Expences";
+// import Register from "../../components/Register/Register";
+// import Login from "../../components/Login/Login";
+// import Info from "../../components/Profile/Info"; // change name
+
+const Loading = () => (
+  <div style={styles.loading}>
+    <CircularProgress size={80} color="primary" />
+  </div>
+);
+
+const Home = Loadable({
+  loader: () => import("../Expences/Expences"),
+  loading: Loading
+});
+
+const LandingPage = Loadable({
+  loader: () => import("../LandingPage/LandingPage"),
+  loading: Loading
+});
+
+const Login = Loadable({
+  loader: () => import("../../components/Login/Login"),
+  loading: Loading
+});
+
+const Register = Loadable({
+  loader: () => import("../../components/Register/Register"),
+  loading: Loading
+});
+
+const Profile = Loadable({
+  loader: () => import("../../components/Profile/Info"),
+  loading: Loading
+});
+
+const styles = {
+  loading: {
+    position: "absolute",
+    top: 0,
+    left: 0,
+    bottom: 0,
+    right: 0,
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    zIndex: "99999",
+    backgroundColor: "rgba(0,0,0, 0.5)"
+  }
+};
 
 class Main extends React.Component {
   render() {
@@ -58,7 +107,7 @@ class Main extends React.Component {
           exact
           path="/home"
           render={props => (
-            <Expences
+            <Home
               {...props}
               isSignedIn={this.props.isSignedIn}
               user={this.props.user}
@@ -69,7 +118,7 @@ class Main extends React.Component {
           exact
           path="/profile"
           render={props => (
-            <Info
+            <Profile
               {...props}
               isSignedIn={this.props.isSignedIn}
               name={this.props.user.name}
